Second Home Work
Locators updated.
Logger added.
Comments removed.
Code refactored.
Class renamed.
